// src/utils/enums/enums.ts
export enum RequestStatusType {
  idle,
  loading,
  succeeded,
  failed,
}

export enum MenuItemsTypes {
  HOME = "home",
  ABOUT = 'about',
  HOUSES = 'houses',
  KITCHEN = 'kitchen',
  ENTERTAINMENTS = 'entertainments',
  NEARESTS = 'nearests',
  SPECIAL = 'special',
  RULES = 'rules',
  GALLERY = 'gallery',
  CONTACTS = 'contacts'
}